# docs/api/catalog-service.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "Kitap kataloğu: herkese açık okuma, admin yönetimi ve servisler arası stok rezervasyonu.\n\n**Erişim:**\n- Herkese açık (`GET /api/books/**`, `GET /api/categories/**`): kimlik gerekmez; `Authorization` başlığı gönderilse de yok sayılır.\n- Admin (`/api/admin/**`): user-service'ten alınan JWT, `Authorization: Bearer <accessToken>`, `role` = `ADMIN` (`bearerAuth`).\n- Internal (`/internal/**`): yalnızca servisler arası, `X-Internal-Api-Key` başlığı (`internalApiKey`). Kullanıcı JWT'si (ADMIN dahil) bu uçlarda geçersizdir. Ayrıntılar: `docs/api/catalog-internal-stock.md`.\n\n**Hatalar:** Tüm hatalar RFC 9457 ProblemDetail olarak `application/problem+json` ile döner (`Problem` şeması). İstemci hatayı `code` alanına göre ayırt eder; doğrulama hatalarında `errors` dizisi alan bazında mesaj içerir. Stok rezervasyonu çakışmalarında ek olarak `bookIds` döner (`StockUnavailableProblem`).\n",
    "title": "Kitap Sepeti Catalog API",
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "/",
      "description": "Dokümanın sunulduğu sunucu"
    }
  ],
...
  "paths": {
...
    "/api/admin/publishers": {
      "get": {
        "description": "Ada göre sıralı.",
        "operationId": "listPublishers",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponsePublisherResponse"
                }
              }
            },
            "description": "Sayfa."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED` (alan hataları `errors` dizisinde) veya `MALFORMED_REQUEST` (okunamayan JSON ya da bozuk path değişkeni)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHORIZED`: token yok, geçersiz ya da süresi dolmuş.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Token yoksa `Bearer`, geçersizse `Bearer error=\"invalid_token\"`.",
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`FORBIDDEN`: token geçerli ama rol `ADMIN` değil."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`INTERNAL_ERROR`: beklenmeyen hata; ayrıntı yanıtta yer almaz."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`AUTHENTICATION_UNAVAILABLE`: token doğrulanamadı çünkü user-service JWKS ucuna ulaşılamıyor; daha sonra tekrar deneyin."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Yayınevlerini listele",
        "tags": [
          "Admin – Publishers"
        ]
      },
      "post": {
        "description": "`slug` verilmezse addan üretilir (Türkçe karakterler sadeleştirilir).",
        "operationId": "createPublisher",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePublisherRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherResponse"
                }
              }
            },
            "description": "Oluşturuldu.",
            "headers": {
              "Location": {
                "description": "Yeni yayınevinin yolu.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                },
                "style": "simple"
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED` (alan hataları `errors` dizisinde) veya `MALFORMED_REQUEST` (okunamayan JSON ya da bozuk path değişkeni)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHORIZED`: token yok, geçersiz ya da süresi dolmuş.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Token yoksa `Bearer`, geçersizse `Bearer error=\"invalid_token\"`.",
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`FORBIDDEN`: token geçerli ama rol `ADMIN` değil."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`SLUG_ALREADY_EXISTS`: slug kullanımda."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`INTERNAL_ERROR`: beklenmeyen hata; ayrıntı yanıtta yer almaz."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`AUTHENTICATION_UNAVAILABLE`: token doğrulanamadı çünkü user-service JWKS ucuna ulaşılamıyor; daha sonra tekrar deneyin."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Yayınevi oluştur",
        "tags": [
          "Admin – Publishers"
        ]
      }
    },
    "/api/admin/publishers/{id}": {
      "delete": {
        "operationId": "deletePublisher",
        "parameters": [
          {
            "description": "Yayınevi id'si",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Silindi."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED` (alan hataları `errors` dizisinde) veya `MALFORMED_REQUEST` (okunamayan JSON ya da bozuk path değişkeni)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHORIZED`: token yok, geçersiz ya da süresi dolmuş.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Token yoksa `Bearer`, geçersizse `Bearer error=\"invalid_token\"`.",
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`FORBIDDEN`: token geçerli ama rol `ADMIN` değil."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RESOURCE_NOT_FOUND`: kayıt yok."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RESOURCE_IN_USE`: yayınevine bağlı kitap var."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`INTERNAL_ERROR`: beklenmeyen hata; ayrıntı yanıtta yer almaz."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`AUTHENTICATION_UNAVAILABLE`: token doğrulanamadı çünkü user-service JWKS ucuna ulaşılamıyor; daha sonra tekrar deneyin."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Yayınevi sil",
        "tags": [
          "Admin – Publishers"
        ]
      },
      "get": {
        "operationId": "getPublisher",
        "parameters": [
          {
            "description": "Yayınevi id'si",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherResponse"
                }
              }
            },
            "description": "Yayınevi."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED` (alan hataları `errors` dizisinde) veya `MALFORMED_REQUEST` (okunamayan JSON ya da bozuk path değişkeni)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHORIZED`: token yok, geçersiz ya da süresi dolmuş.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Token yoksa `Bearer`, geçersizse `Bearer error=\"invalid_token\"`.",
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`FORBIDDEN`: token geçerli ama rol `ADMIN` değil."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RESOURCE_NOT_FOUND`: kayıt yok."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`INTERNAL_ERROR`: beklenmeyen hata; ayrıntı yanıtta yer almaz."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`AUTHENTICATION_UNAVAILABLE`: token doğrulanamadı çünkü user-service JWKS ucuna ulaşılamıyor; daha sonra tekrar deneyin."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Yayınevi getir",
        "tags": [
          "Admin – Publishers"
        ]
      },
      "patch": {
        "description": "Kısmi güncelleme; yalnızca ad değişirse slug korunur.",
        "operationId": "updatePublisher",
        "parameters": [
          {
            "description": "Yayınevi id'si",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePublisherRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherResponse"
                }
              }
            },
            "description": "Güncellenmiş yayınevi."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED` (alan hataları `errors` dizisinde) veya `MALFORMED_REQUEST` (okunamayan JSON ya da bozuk path değişkeni)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHORIZED`: token yok, geçersiz ya da süresi dolmuş.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Token yoksa `Bearer`, geçersizse `Bearer error=\"invalid_token\"`.",
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`FORBIDDEN`: token geçerli ama rol `ADMIN` değil."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RESOURCE_NOT_FOUND`: kayıt yok."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`SLUG_ALREADY_EXISTS`: slug kullanımda."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`INTERNAL_ERROR`: beklenmeyen hata; ayrıntı yanıtta yer almaz."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`AUTHENTICATION_UNAVAILABLE`: token doğrulanamadı çünkü user-service JWKS ucuna ulaşılamıyor; daha sonra tekrar deneyin."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Yayınevi güncelle",
        "tags": [
          "Admin – Publishers"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePublisherRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
          }
        },
        "required": [
          "name"
        ]
      },
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Hatalı alanın adı.",
            "example": "slug"
          },
          "message": {
            "type": "string",
            "description": "Doğrulama mesajı.",
            "example": "must not be blank"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
...
      "PageResponsePublisherResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublisherResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Problem": {
        "type": "object",
        "description": "RFC 9457 Problem Details. İstemci hatayı `code` alanına göre ayırt etmelidir; `title`/`detail` insan içindir ve değişebilir.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Makine tarafından okunacak hata kodu.",
            "enum": [
              "VALIDATION_FAILED",
              "MALFORMED_REQUEST",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "RESOURCE_NOT_FOUND",
              "METHOD_NOT_ALLOWED",
              "NOT_ACCEPTABLE",
              "UNSUPPORTED_MEDIA_TYPE",
              "CONFLICT",
              "SLUG_ALREADY_EXISTS",
              "ISBN_ALREADY_EXISTS",
              "RESOURCE_IN_USE",
              "CONCURRENT_MODIFICATION",
              "CATEGORY_CYCLE",
              "BOOK_NOT_PUBLISHABLE",
              "STOCK_BELOW_RESERVED",
              "INSUFFICIENT_STOCK",
              "BOOK_NOT_AVAILABLE",
              "RESERVATION_MISMATCH",
              "RESERVATION_RELEASED",
              "RESERVATION_COMMITTED",
              "AUTHENTICATION_UNAVAILABLE",
              "INTERNAL_ERROR"
            ],
            "example": "SLUG_ALREADY_EXISTS"
          },
          "detail": {
            "type": "string",
            "description": "Genel açıklama; kullanıcı verisi içermez.",
            "example": "Slug is already in use."
          },
          "errors": {
            "type": "array",
            "description": "Yalnızca `VALIDATION_FAILED`'da: alan bazında hatalar. Gönderilen değer yer almaz.",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "instance": {
            "type": "string",
            "format": "uri-reference",
            "description": "İsteğin yolu.",
            "example": "/api/admin/publishers"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP durum kodu.",
            "example": 409
          },
          "title": {
            "type": "string",
            "description": "HTTP durumunun kısa adı.",
            "example": "Conflict"
          },
          "type": {
            "type": "string",
            "format": "uri",
            "description": "Problem tipi; varsayılan `about:blank` olduğunda yanıtta yer almaz."
          }
        },
        "required": [
          "code",
          "instance",
          "status",
          "title"
        ]
      },
...
      "PublisherResponse": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdatePublisherRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0,
            "pattern": ".*\\S.*"
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "user-service `POST /api/auth/login` yanıtındaki `accessToken`; `role` = `ADMIN` olmalı.",
        "scheme": "bearer",
        "type": "http"
      },
      "internalApiKey": {
        "description": "Servisler arası ham anahtar (yalnızca çağıran serviste saklanır; catalog yalnızca SHA-256 özetini bilir). Kullanıcı JWT'si bu uçlarda geçersizdir.",
        "in": "header",
        "name": "X-Internal-Api-Key",
        "type": "apiKey"
      }
    }
  }
}